% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roci_functions.R
\name{auc_roci}
\alias{auc_roci}
\title{Compute area under the curve (AUC) for curve provided by ROCI algorithm.}
\usage{
auc_roci(roci_results)
}
\value{
A numeric giving the AUC value.
}
\description{
Compute area under the curve (AUC) for curve computed via \code{roci()} function using the maximal sensitivity corresponding to each specificity.
}
\examples{
set.seed(123)
ph_values <- rnorm(20, 7, 2)
med_class <- ifelse(ph_values >= quantile(ph_values, 0.25) &
  ph_values <= quantile(ph_values, 0.75),
sample(c(0, 1), prob = c(0.3, 0.7)),
sample(c(0, 1), prob = c(0.8, 0.2))
)
roci_results <- roci(ph_values, med_class)
auc_roci(roci_results)
}
