% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roci_functions.R
\name{plot_roci}
\alias{plot_roci}
\title{Plot results of ROCI algorithm}
\usage{
plot_roci(
  roci_results,
  hide_whiskers = FALSE,
  add_width_plot = TRUE,
  add_yi_plot = TRUE,
  add_cutoff_plot = TRUE,
  add_auc_plot = TRUE,
  ...
)
}
\arguments{
\item{roci_results}{A data frame as returned by \code{roci()}.}

\item{add_width_plot}{Logical if width of intervals should be added to plot.}

\item{add_yi_plot}{Logical if maximal Youden-Index should be written to plot
and if the corresponding interval should be highlighted.}

\item{add_cutoff_plot}{Logical if corresponding interval of maximal
Youden-Index should be written to plot.}

\item{add_auc_plot}{Logical if corresponding AUC should be written to plot.}

\item{...}{Additional arguments passed to ggplot.}

\item{hide_whisker}{Logical if suboptimal sensitivity-specificity pairs for
fixed interval lengths should be hidden in the plot.}
}
\value{
A ggplot.
}
\description{
Plot a ROC curve for intervals as calculated by \code{roci()}.
}
\examples{
set.seed(123)
ph_values <- rnorm(20, 7, 2)
med_class <- ifelse(ph_values >= quantile(ph_values, 0.25) &
  ph_values <= quantile(ph_values, 0.75),
sample(c(0, 1), prob = c(0.3, 0.7)),
sample(c(0, 1), prob = c(0.8, 0.2))
)
roci_results <- roci(ph_values, med_class)
plot_roci(roci_results)
}
