% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCI-package.R
\docType{package}
\name{ROCI-package}
\alias{ROCI}
\alias{ROCI-package}
\title{ROCI: ROCI package}
\description{
Provides a set of functions to calculate and visualize the Receiver Operating Characteristic (ROC) curve, the Area Under the Curve (AUC) and other statistical metrics for a binary classification model, where the two classes are defined by an interval i.e. two thresholds.
}
\author{
\strong{Maintainer}: Dominik Thiele \email{dthiele@tcc-clinicalsolutions.de} (\href{https://orcid.org/0000-0003-1073-0747}{ORCID})

}
\keyword{internal}
