% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roci_functions.R
\name{roci}
\alias{roci}
\title{Calculates the ROC for intervals}
\usage{
roci(data, classes)
}
\arguments{
\item{data}{A numerical vector, which should be classified into a binary.
outcome.}

\item{classes}{A numerical binary vector with the knwon classes of the data.}
}
\value{
A data.frame containing the cut-off values, sensitivity, specificity
and the Youden-Index for all possible intervals.
}
\description{
\code{roci()} calculates sensitivity, specificity and the Youden-Index for tests,
where two cut-offs (interval) are used for binary classification.
All possible intervals are tested. A full description of the algorithm is
available in Krannich 2023 ...
}
\examples{
set.seed(123)
ph_values <- rnorm(20, 7, 2)
med_class <- ifelse(ph_values >= quantile(ph_values, 0.25) &
  ph_values >= quantile(ph_values, 0.75),
sample(c(0, 1), prob = c(0.3, 0.7)),
sample(c(0, 1), prob = c(0.8, 0.2))
)
roci(ph_values, med_class)
}
